requirements.txt
asgiref==3.8.1
Django==5.2.3
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.2
django-crispy-forms
crispy-bootstrap4
django-allauth
django-allauth[socialaccount]
environs[django]
pillow
django-debug-toolbar
whitenoise
gunicorn